services/analysis/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
Pillow>=11.0.0
piexif>=1.1.3
imagehash>=4.3.1
scikit-image>=0.24.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
scipy>=1.14.0
c2pa-python>=0.28.0
requests>=2.32.0
beautifulsoup4>=4.12.0
huggingface-hub>=0.30.0
python-dotenv>=1.0.1
